requirements.txt
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.10.0

pyyaml>=6.0.0
requests>=2.31.0
